\name{AffySpikeU95Subset}
\alias{AffySpikeU95Subset}
\alias{HGU95subset1016}
\alias{spikedProbesU95}
\title{Spike-in data}
\description{
This \code{\link[affy]{AffyBatch-class}} represents part of the Affymetrix spike-in data set.
}
\usage{data(AffySpikeU95Subset)}
\format{An \code{\link[affy]{AffyBatch-class}} containing 6 arrays.}
\source{
Array 1, 2, 21, 22, 41, and 42 of the Affymetrix U95 Latin square data set, thus the 6 arrays of group A and B, and 
a random subset of 1000 probe-sets together with the 16 spike-in probe-sets. 
}
\keyword{datasets}
