\name{plw-package}
\alias{plw-package}
\docType{package}
\title{plw}
\description{
Locally moderated weighted analysis of microarray data, at the probe level (PLW) or at the level of expression indexes (LMW).
}
\author{Magnus \eqn{\mbox{\AA}}{A}strand}
\references{
\eqn{\mbox{\AA}}{A}strand, M. et al. (2007a). Improved covariance matrix estimators for weighted analysis of microarray data. Journal of Computational Biology, Accepted.

\eqn{\mbox{\AA}}{A}strand, M. et al. (2007b). Empirical Bayes models for multiple-probe type arrays at the probe level. Bioinformatics, Submitted 1 October 2007.
}
\keyword{package}
\seealso{plw, lwm}
